Refs #17304. Separate LoadBankFromDisk to another file.
Speed test is fine.
